chat-service/requirements.txt
blinker==1.9.0
click==8.1.8
Flask==3.1.0
Flask-SQLAlchemy==3.1.1
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
psycopg2-binary==2.9.10
SQLAlchemy==2.0.40
typing_extensions==4.13.1
Werkzeug==3.1.3
psycopg2-binary
PyJWT
python-dotenv==1.0.0
psycopg2-pool==1.1
gunicorn==20.1.0